Update README.md
Attempting to change the voice of the README file so that it is inspired by the original effort and not to appear as an endorsement. Also trying to correct the lazily copied text from the original project description. Hopefully this is a better start to explaining my motivations to work on this project and provide this tool to conference attendees worldwide.